staking fix for 1.1.3

+++ chia/consensus/pot_iterations.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from chia.consensus.constants import ConsensusConstants
 from chia.consensus.pos_quality import _expected_plot_size
 from chia.types.blockchain_format.sized_bytes import bytes32
@@ -48,7 +50,7 @@ def calculate_iterations_quality(
     quality_string: bytes32,
     size: int,
     difficulty: uint64,
-    difficulty_coeff: float,
+    difficulty_coeff: Decimal,
     cc_sp_output_hash: bytes32,
 ) -> uint64:
     """
